feat(PN-13919): move model in NewNotification.ts
