+++ packages/pn-pa-webapp/src/components/NewNotification/__test__/PreliminaryInformationsContent.test.tsx
@@ -9,8 +9,7 @@ import {
   testFormElements,
 } from '@pagopa-pn/pn-commons/src/test-utils';
 
-import { NewNotificationLangOther, PaymentModel } from '../../../models/NewNotification';
-import { PreliminaryInformationsPayload } from '../../../redux/newNotification/types';
+import { NewNotificationLangOther, PaymentModel, PreliminaryInformationsPayload } from '../../../models/NewNotification';
 import PreliminaryInformationsContent from '../PreliminaryInformationsContent';
 
 // mock imports
